+++ R/get_peaks.R
@@ -9,6 +9,9 @@
 #' function is fit to the signal using \code{\link{fit_peaks}}). The area is then
 #' calculated using a trapezoidal approximation.
 #' 
+#' The \code{sd}, \code{FWHM}, \code{tau}, and \code{area} are returned in units
+#' determined by \code{time.units}. By defaults the units are in minutes.
+#' 
 #' @aliases get_peaks
 #' @importFrom stats median
 #' @param chrom_list A list of profile matrices, each of the same dimensions
@@ -20,6 +23,8 @@
 #' @param sd.max Maximum width (standard deviation) for peaks. Defaults to 50.
 #' @param max.iter Maximum number of iterations for non-linear least squares
 #' in \code{\link{fit_peaks}}.
+#' @param time.units Units of \code{sd}, \code{FWHM}, \code{area}, and \code{tau}
+#' (if applicable). Options are milliseconds \code{"ms"} or minutes \code{"min"}.
 #' @param \dots Additional arguments to \code{\link{find_peaks}}.
 #' @return The result is an S3 object of class \code{peak_list}, containing a nested
 #' list of data.frames containing information about the peaks fitted for each
@@ -55,7 +60,9 @@
 #' @export get_peaks
 
 get_peaks <- function (chrom_list, lambdas, fit = c("egh", "gaussian", "raw"),
-                       sd.max=50, max.iter=100, ...){
+                       sd.max=50, max.iter=100, time.units = c("min", "ms"), ...){
+  time.units <- match.arg(time.units, c("min", "ms"))
+  tfac <- switch(time.units, "ms" = 60000, "min" = 1)
   fit <- match.arg(fit, c("egh", "gaussian", "raw"))
   if (class(chrom_list)[1] == "matrix")
     chrom_list <- list(chrom_list)
@@ -94,14 +101,15 @@ get_peaks <- function (chrom_list, lambdas, fit = c("egh", "gaussian", "raw"),
     x <- lambda
     x[, c('rt', 'start', 'end')] <- sapply(c('rt', 'start', 'end'),
                                            function(j) timepoints[x[,j]])
-    x[, c('sd', 'FWHM')] <- x[, c('sd', 'FWHM')] * tdiff
-    if (!is.null(x$tau)){x[, c('tau')] <- x[, c('tau')] * tdiff} 
+    x[, c('sd', 'FWHM', 'area')] <- x[, c('sd', 'FWHM', 'area')] * tdiff * tfac
+    if (!is.null(x$tau)){x[, c('tau')] <- x[, c('tau')] * tdiff * tfac} 
     x
   }))
   structure(result,
             chrom_list = chrom_list_str,
             lambdas = deparse(substitute(lambdas)), fit=fit, sd.max=sd.max,
             max.iter = max.iter,
+            time.units = time.units,
             class = "peak_list")
 }
 
@@ -137,6 +145,8 @@ get_peaks <- function (chrom_list, lambdas, fit = c("egh", "gaussian", "raw"),
 plot.peak_list <- function(x, ..., chrom_list=NULL, index=1, lambda=NULL,
                        points=FALSE, ticks=FALSE, a=0.5, color=NULL,
                        cex.points=0.5){
+  time.units <- attributes(x)$time.units
+  tfac <- switch(time.units, "min" = 1, "ms" = 1/60000)
   if (is.null(chrom_list)){
     chrom_list <- get_chrom_list(x)
   }
@@ -174,13 +184,14 @@ plot.peak_list <- function(x, ..., chrom_list=NULL, index=1, lambda=NULL,
     peak.loc<-seq.int((pks$start[i]),(pks$end[i]), by = res)
       if (fit == "gaussian"){
         yvals <- gaussian(peak.loc, center=pks$rt[i],
-                          width=pks$sd[i],height = pks$height[i])
+                          width=pks$sd[i]*tfac, height = pks$height[i])
         if (is.null(color))
           color <- "red"
       }
       else if (fit == "egh"){
         yvals <- egh(x = peak.loc, center = pks$rt[i],
-                     width=pks$sd[i], height = pks$height[i], tau=pks$tau[i])
+                     width=pks$sd[i]*tfac, height = pks$height[i],
+                     tau=pks$tau[i]*tfac)
         if (is.null(color))
           color <- "purple"
       }